# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Garoua Explorer API",
    "version": "1.0.0",
    "description": "API REST de l'application de recommandation touristique Garoua Explorer."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Développement local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "TransportOption": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "type": "string",
            "enum": [
              "MOTO_TAXI",
              "BUS",
              "AGENCY",
              "CAR_RENTAL",
              "AIRPORT"
            ]
          },
          "name": {
            "type": "string"
          },
          "basePrice": {
            "type": "number"
          },
          "priceUnit": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/transport": {
      "get": {
        "summary": "Lister les options de transport",
        "tags": [
          "Transport"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransportOption"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Créer une option de transport (Admin)",
        "tags": [
          "Transport"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/transport/{id}": {
      "patch": {
        "summary": "Modifier les prix (Admin)",
        "tags": [
          "Transport"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "summary": "Supprimer une option (Admin)",
        "tags": [
          "Transport"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "summary": "Lister les utilisateurs (Admin)",
        "tags": [
          "Utilisateurs & Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Voir un profil",
        "tags": [
          "Utilisateurs & Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "summary": "Modifier un profil (soi-même, ou Admin)",
        "tags": [
          "Utilisateurs & Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "summary": "Supprimer un compte (Admin)",
        "tags": [
          "Utilisateurs & Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}